Test hash function, Change some names for the functions for readability

+++ src/hash/ft_hash_algo.c
@@ -6,7 +6,7 @@
 /*   By: sebasnadu <johnavar@student.42berlin.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/09 23:08:55 by sebasnadu         #+#    #+#             */
-/*   Updated: 2024/02/10 21:56:27 by sebasnadu        ###   ########.fr       */
+/*   Updated: 2024/02/11 12:13:10 by sebasnadu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,16 +28,16 @@
 } */
 
 // NOTE: This function is an implementation of the djb2 hashing algorithm
-unsigned long int	ft_hash_algo(char *key, unsigned int size)
+unsigned long	ft_hash_algo(char *key, unsigned int size)
 {
 	unsigned long	hash;
-	unsigned int	c;
+	unsigned long	c;
 
 	hash = 14695981039346656037UL;
 	while (*key)
 	{
 		c = *key++;
-		hash ^= (unsigned long)c;
+		hash ^= c;
 		hash *= 1099511628211UL;
 	}
 	return (hash % size);
